Drawing battery widget, compass widget and accuracy radius; overall improvements

+++ src/core/tile.h
@@ -15,7 +15,7 @@ public:
 
   static std::string getTileKey(uint32_t x, uint32_t y, uint32_t z);
 
-  static std::pair<double, double> convertLatLongToTileXY(float latitude, float longitude, uint8_t zoom);
+  static std::pair<double, double> convertLatLongToTileXY(double latitude, double longitude, uint8_t zoom);
 
   const std::string key;
   const uint32_t x;
